Added loading only on first time playing the game, doesnt preload on main page anymore
