+++ src/hooks/useGame.ts
@@ -1,7 +1,7 @@
 import { setCookie } from 'cookies-next';
+import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { updateLevelsAnswers } from '../services/level';
-import { Alert } from '../types/alert';
 import { Level, LEVEL_TYPE_ENUM, PreviousAnswerLevel } from '../types/level';
 
 const useGame = (
@@ -9,9 +9,9 @@ const useGame = (
   highestScoreCookie: number,
   setFinishedGameModal: (_v: boolean) => void
 ) => {
-  const [gameStarted, setGameStarted] = useState(false);
+  const router = useRouter();
+
   const [currentScore, setCurrentScore] = useState(0);
-  const [alert, setAlert] = useState<Alert>({});
 
   const [previousAnswers, setPreviousAnswers] = useState<PreviousAnswerLevel[]>(
     []
@@ -52,9 +52,9 @@ const useGame = (
   };
 
   const endGame = () => {
+    router.push('/');
     if (currentScore > highestScoreCookie) updateHighScore();
     setCurrentScore(0);
-    setGameStarted(false);
 
     updateAnswers(previousAnswers);
     setCurrentLevel(getRandomLevel());
@@ -85,10 +85,6 @@ const useGame = (
   };
 
   return {
-    alert,
-    setAlert,
-    gameStarted,
-    setGameStarted,
     currentScore,
     setCurrentScore,
     currentLevel,
