+++ src/pages/index.tsx
@@ -1,124 +1,53 @@
-import Image from 'next/image';
+import Link from 'next/link';
 import type { GetServerSideProps, NextPage } from 'next';
 import { Level } from '../types/level';
-import ReactLoading from 'react-loading';
-
-import useGame from '../hooks/useGame';
 
 import Layout from '../components/layout';
-import StartGame from '../components/start-game';
-import Game from '../components/game';
-import EndGameModal from '../components/end-game-modal';
 import { getCookie } from 'cookies-next';
-import { useState } from 'react';
-import { getAllLevels } from '../services/level';
 interface IMain {
   levels: Level[];
   highestScoreCookie: number;
+  hasPreloaded: boolean;
 }
 
-const Main: NextPage<IMain> = ({ levels, highestScoreCookie }) => {
-  const [openFinishGameModal, setOpenFinishGameModal] = useState(false);
-  const [loading, setLoading] = useState(true);
-
-  const {
-    alert,
-    setAlert,
-    gameStarted,
-    setGameStarted,
-    currentScore,
-    currentLevel,
-    highestScoreClientState,
-    previousAnswers,
-    handleAnswer,
-    endGame,
-  } = useGame(levels, highestScoreCookie, setOpenFinishGameModal);
-
-  const preloadImages = () => {
-    return (
-      <>
-        {levels.map((level, idx) => (
-          <div key={level._id}>
-            <Image
-              src={`/assets/${level.image_name}`}
-              alt="Level Image"
-              width={700}
-              height={550}
-              quality={100}
-              objectFit="cover"
-              objectPosition="50% 20%"
-              className="rounded shadow-2xl"
-              layout="responsive"
-              onLoadingComplete={() =>
-                idx + 1 == levels.length ? setLoading(false) : null
-              }
-            />
-          </div>
-        ))}
-      </>
-    );
-  };
-
-  if (loading)
-    return (
-      <>
-        <div className="bg-secondary flex-col w-screen h-screen flex items-center justify-center">
-          <div className="flex flex-col items-center justify-center">
-            <Image
-              src="/assets/ai-or-human-logo-updated.png"
-              alt="AI or HUMAN Logo with an icon of a robot and a human"
-              quality={100}
-              width={400}
-              height={200}
-              priority
-            />
-
-            <ReactLoading type="spinningBubbles" color="rgb(255 211 105)" />
-          </div>
-          <p className="text-slate-200 text-xs fixed bottom-1">
-            When opening the first time, loading might take longer.
-          </p>
-          <div className="fixed overflow-scroll overflow-x-hidden">
-            {preloadImages()}
-          </div>
-        </div>
-      </>
-    );
-
+const Main: NextPage<IMain> = ({ highestScoreCookie, hasPreloaded }) => {
   return (
-    <Layout alert={alert} setAlert={setAlert}>
-      <EndGameModal
-        currentScore={currentScore}
-        levels={levels}
-        previousAnswers={previousAnswers}
-        openFinishGameModal={openFinishGameModal}
-        setOpenFinishGameModal={setOpenFinishGameModal}
-        endGameTrigger={endGame}
-      />
-      {!gameStarted ? (
-        <StartGame
-          highestScore={highestScoreClientState}
-          setGameStarted={setGameStarted}
-        />
-      ) : (
-        <Game
-          currentScore={currentScore}
-          currentLevel={currentLevel}
-          handleGameAnswer={handleAnswer}
-        />
-      )}
+    <Layout>
+      <div className="w-3/4 lg:w-2/4 flex flex-col justify-between items-center">
+        <h1 className="text-center text-sm xs:text-xl sm:text-2xl md:text-3xl font-bold">
+          Are the images created by an Human or generated by AI?
+        </h1>
+        <p className="text-center text-xs sm:text-md md:text-xl">
+          As tecnology is evolving, so is the AI capabilities. With the current
+          advancements, it is possible to generate images with a simple text
+          prompt. In this game you are challenged to choose wether an image was
+          created by an Human or virtually generated by an Artifical
+          Inteligence.
+        </p>
+        <h2 className="self-center py-4 text-xs xs:text-sm sm:text-lg md:text-xl">
+          Your Highest Score: {highestScoreCookie}
+        </h2>
+        <Link href={hasPreloaded ? '/game' : '/loading'}>
+          <button
+            type="button"
+            className="w-3/4 sm:w-2/4 inline-block px-2 py-2 xs:px-6 border-2 border-detail text-detail font-medium text-md leading-tight uppercase rounded hover:bg-secondary hover:shadow-lg focus:outline-none focus:ring-0 transition duration-150 ease-in-out"
+          >
+            Play Game
+          </button>
+        </Link>
+      </div>
     </Layout>
   );
 };
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
-  const levels = await getAllLevels();
   const highestScore = getCookie('highestScore', ctx);
+  const hasPreloaded = getCookie('hasPreloaded', ctx);
 
   return {
     props: {
-      levels,
       highestScoreCookie: highestScore || 0,
+      hasPreloaded: hasPreloaded || false,
     },
   };
 };
